Added the ability to override the kube config cluster server.
This change allows users with a private cluster to be able to use SSH
tunnelling to access the cluster from their local machine via a
port other than 443.

The KUBECONFIG_CLUSTER_ENDPOINT environment variable can now be
set to override the cluster endpoint in the kube config object
that's generated via kubeconfig.go.

This change references #1552 and #1618